move component to screens
