+++ src/screens/item-detail-container/index.jsx
@@ -2,10 +2,10 @@ import React, { useEffect, useState } from "react";
 
 import "./styles.css";
 
-import { getProductById } from "../mocks/products-mock";
+import { getProductById } from "../../mocks/products-mock";
 
-import ItemDetail from "../components/item-detail";
-import Loading from "../components/loading";
+import ItemDetail from "../../components/item-detail";
+import Loading from "../../components/loading";
 
 function ItemDetailContainer() {
   const [loading, setLoading] = useState(true);
@@ -17,10 +17,10 @@ function ItemDetailContainer() {
     });
   }, []);
   return (
-    <>
+    <div className="container">
       {loading && <Loading />}
       {!loading && <ItemDetail item={item} />}
-    </>
+    </div>
   );
 }
