+++ src/screens/item-detail-container/styles.css
@@ -1,3 +1,1 @@
-.container-item-detail {
-  margin: 25px 0px;
-}
+
