--- src/screens/shop/index.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-
-import "./style.css";
-
-import ItemListContainer from "../../components/item-list-container";
-
-function Shop() {
-  return (
-    <div className="container">
-      <ItemListContainer />
-    </div>
-  );
-}
-
-export default Shop;
